@@ -3,48 +3,60 @@ package subscribe
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
 	"reflect"
 
 	"gopkg.in/yaml.v2"
+
+	"github.com/icpd/subscribe2clash/internal/acl"
 )
 
-var (
-	OutputFile string
-)
-
-func (c *Clash) LoadTemplate(path string, proxies []interface{}) []byte {
-	_, err := os.Stat(path)
-	if err != nil && os.IsNotExist(err) {
-		log.Printf("[%s] template doesn't exist.", path)
-		return nil
-	}
-	buf, err := ioutil.ReadFile(path)
-	if err != nil {
-		log.Printf("[%s] template open the failure.", path)
-		return nil
-	}
-	err = yaml.Unmarshal(buf, &c)
-	if err != nil {
-		log.Printf("[%s] Template format error.", path)
+func (c *Clash) LoadTemplate() []byte {
+	if !c.NodeOnly() {
+		_, err := os.Stat(c.path)
+		if err != nil && os.IsNotExist(err) {
+			log.Printf("[%s] template doesn't exist. err: %v", c.path, err)
+			return nil
+		}
+		buf, err := os.ReadFile(c.path)
+		if err != nil {
+			log.Printf("[%s] template open the failure. err: %v", c.path, err)
+			return nil
+		}
+		err = yaml.Unmarshal(buf, &c)
+		if err != nil {
+			log.Printf("[%s] Template format error. err: %v", c.path, err)
+		}
 	}
 
-	c.Proxy = nil
-
-	var proxy []map[string]interface{}
-	var proxiesStr []string
+	var proxiesName []string
 	names := map[string]int{}
 
-	for _, proto := range proxies {
-		o := reflect.ValueOf(proto)
-		nameField := o.FieldByName("Name")
-		proxyItem := make(map[string]interface{})
-		j, _ := json.Marshal(proto)
-		_ = json.Unmarshal(j, &proxyItem)
+	for _, proto := range c.rawProxies {
+		proxy := make(map[string]any)
+		p, ok := proto.(map[string]any)
+		if ok {
+			proxy = p
+		} else {
+			j, err := json.Marshal(proto)
+			if err != nil {
+				log.Printf("json marshal error: %v", err)
+				continue
+			}
+			_ = json.Unmarshal(j, &proxy)
+		}
 
-		name := nameField.String()
+		var name string
+		switch reflect.TypeOf(proto).Kind() {
+		case reflect.Struct:
+			name = reflect.ValueOf(proto).FieldByName("Name").String()
+		case reflect.Map:
+			if v, ok := proto.(map[string]any); ok {
+				name = v["name"].(string)
+			}
+		}
+
 		if index, ok := names[name]; ok {
 			names[name] = index + 1
 			name = fmt.Sprintf("%s%d", name, index+1)
@@ -52,35 +64,45 @@ func (c *Clash) LoadTemplate(path string, proxies []interface{}) []byte {
 			names[name] = 0
 		}
 
-		proxyItem["name"] = name
-		proxy = append(proxy, proxyItem)
-		c.Proxy = append(c.Proxy, proxyItem)
-		proxiesStr = append(proxiesStr, name)
+		proxy["name"] = name
+		c.Proxies = append(c.Proxies, proxy)
+		proxiesName = append(proxiesName, name)
 	}
 
-	c.Proxy = proxy
-
-	for _, group := range c.ProxyGroup {
-		groupProxies, ok := group["proxies"].([]interface{})
+	for _, group := range c.ProxyGroups {
+		groupProxies, ok := group["proxies"].([]any)
 		if !ok {
 			continue
 		}
 		for i, p := range groupProxies {
 			if p == "1" {
 				groupProxies = groupProxies[:i]
-				var tmpGroupProxies []string
+
+				var groupProxiesName []string
 				for _, s := range groupProxies {
-					tmpGroupProxies = append(tmpGroupProxies, s.(string))
+					groupProxiesName = append(groupProxiesName, s.(string))
 				}
-				tmpGroupProxies = append(tmpGroupProxies, proxiesStr...)
-				group["proxies"] = tmpGroupProxies
+				groupProxiesName = append(groupProxiesName, proxiesName...)
+
+				group["proxies"] = groupProxiesName
 				break
 			}
 		}
 
 	}
 
-	d, err := yaml.Marshal(c)
+	var yamlOut any = c
+	if c.NodeOnly() {
+		type nodeOnly struct {
+			Proxies []map[string]any `yaml:"proxies"`
+		}
+
+		yamlOut = nodeOnly{
+			Proxies: c.Proxies,
+		}
+	}
+
+	d, err := yaml.Marshal(yamlOut)
 	if err != nil {
 		return nil
 	}
@@ -88,10 +110,18 @@ func (c *Clash) LoadTemplate(path string, proxies []interface{}) []byte {
 	return d
 }
 
-func GenerateClashConfig(proxies []interface{}) ([]byte, error) {
-	clash := Clash{}
+func (c *Clash) NodeOnly() bool {
+	return c.nodeOnly
+}
 
-	r := clash.LoadTemplate(OutputFile, proxies)
+func GenerateClashConfig(proxies []any, nodeOnly bool) ([]byte, error) {
+	clash := Clash{
+		path:       acl.GlobalGen.OutputFile,
+		rawProxies: proxies,
+		nodeOnly:   nodeOnly,
+	}
+
+	r := clash.LoadTemplate()
 	if r == nil {
 		return nil, fmt.Errorf("sublink 返回数据格式不对")
 	}
