@@ -8,25 +8,45 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/whoisix/subscribe2clash/internal/xbase64"
+	"github.com/icpd/subscribe2clash/internal/xbase64"
+	"github.com/spf13/cast"
+	"gopkg.in/yaml.v2"
+)
+
+const (
+	ssrHeader      = "ssr://"
+	vmessHeader    = "vmess://"
+	ssHeader       = "ss://"
+	trojanHeader   = "trojan://"
+	hysteriaHeader = "hysteria://"
 )
 
 var (
 	//ssReg      = regexp.MustCompile(`(?m)ss://(\w+)@([^:]+):(\d+)\?plugin=([^;]+);\w+=(\w+)(?:;obfs-host=)?([^#]+)?#(.+)`)
-	ssReg2 = regexp.MustCompile(`(?m)ss://([\-0-9a-z]+):(.+)@(.+):(\d+)(.+)?#(.+)`)
+	ssReg2 = regexp.MustCompile(`(?m)([\-0-9a-z]+):(.+)@(.+):(\d+)(.+)?#(.+)`)
+	ssReg  = regexp.MustCompile(`(?m)([^@]+)(@.+)?#?(.+)?`)
 
 	trojanReg  = regexp.MustCompile(`(?m)^trojan://(.+)@(.+):(\d+)\?allowInsecure=\d&peer=(.+)#(.+)`)
 	trojanReg2 = regexp.MustCompile(`(?m)^trojan://(.+)@(.+):(\d+)#(.+)$`)
 )
 
-func ParseProxy(contentSlice []string) []interface{} {
-	var proxies []interface{}
+func ParseProxy(contentSlice []string) []any {
+	var proxies []any
 	for _, v := range contentSlice {
+		// try unmarshal clash config
+		var c Clash
+		if err := yaml.Unmarshal([]byte(v), &c); err == nil {
+			for _, pg := range c.Proxies {
+				proxies = append(proxies, pg)
+			}
+			continue
+		}
+
 		// ssd
 		if strings.Contains(v, "airport") {
 			ssSlice := ssdConf(v)
 			for _, ss := range ssSlice {
-				if !filterNode(ss.Name) {
+				if ss.Name != "" {
 					proxies = append(proxies, ss)
 				}
 			}
@@ -35,39 +55,95 @@ func ParseProxy(contentSlice []string) []interface{} {
 
 		scanner := bufio.NewScanner(strings.NewReader(v))
 		for scanner.Scan() {
-			switch {
-			case strings.HasPrefix(scanner.Text(), "ssr://"):
-				s := scanner.Text()[6:]
-				s = strings.TrimSpace(s)
-				ssr := ssrConf(s)
-				if ssr.Name != "" && !filterNode(ssr.Name) {
-					proxies = append(proxies, ssr)
-				}
-			case strings.HasPrefix(scanner.Text(), "vmess://"):
-				s := scanner.Text()[8:]
-				s = strings.TrimSpace(s)
-				clashVmess := v2rConf(s)
-				if clashVmess.Name != "" && !filterNode(clashVmess.Name) {
-					proxies = append(proxies, clashVmess)
-				}
-			case strings.HasPrefix(scanner.Text(), "ss://"):
-				s := strings.TrimSpace(scanner.Text())
-				ss := ssConf(s)
-				if ss.Name != "" && !filterNode(ss.Name) {
-					proxies = append(proxies, ss)
-				}
-			case strings.HasPrefix(scanner.Text(), "trojan://"):
-				s := scanner.Text()
-				s = strings.TrimSpace(s)
-				trojan := trojanConf(s)
-				if trojan.Name != "" && !filterNode(trojan.Name) {
-					proxies = append(proxies, trojan)
-				}
+			proxy := parseProxy(scanner.Text())
+			if proxy != nil {
+				proxies = append(proxies, proxy)
 			}
+		}
+
+		if err := scanner.Err(); err != nil {
+			log.Printf("parse proxy failed, err: %v", err)
 		}
 	}
 
 	return proxies
+}
+
+func subProtocolBody(proxy string, prefix string) string {
+	return strings.TrimSpace(proxy[len(prefix):])
+}
+
+func parseProxy(proxy string) any {
+	switch {
+	case strings.HasPrefix(proxy, ssrHeader):
+		return ssrConf(subProtocolBody(proxy, ssrHeader))
+	case strings.HasPrefix(proxy, vmessHeader):
+		return v2rConf(subProtocolBody(proxy, vmessHeader))
+	case strings.HasPrefix(proxy, ssHeader):
+		return ssConf(subProtocolBody(proxy, ssHeader))
+	case strings.HasPrefix(proxy, trojanHeader):
+		return trojanConf(subProtocolBody(proxy, trojanHeader))
+	case strings.HasPrefix(proxy, hysteriaHeader):
+		return hysteriaConf(proxy)
+	}
+
+	return nil
+}
+
+type ClashHysteria struct {
+	Name                string   `yaml:"name"`
+	Type                string   `yaml:"type"`
+	Server              string   `yaml:"server"`
+	Port                int      `yaml:"port"`
+	AuthStr             string   `yaml:"auth-str"`
+	Obfs                string   `yaml:"obfs"`
+	ObfsParams          string   `yaml:"obfs-param"`
+	Alpn                []string `yaml:"alpn"`
+	Protocol            string   `yaml:"protocol"`
+	Up                  string   `yaml:"up"`
+	Down                string   `yaml:"down"`
+	Sni                 string   `yaml:"sni"`
+	SkipCertVerify      bool     `yaml:"skip-cert-verify"`
+	RecvWindowConn      int      `yaml:"recv-window-conn"`
+	RecvWindow          int      `yaml:"recv-window"`
+	Ca                  string   `yaml:"ca"`
+	CaStr               string   `yaml:"ca-str"`
+	DisableMtuDiscovery bool     `yaml:"disable_mtu_discovery"`
+	Fingerprint         string   `yaml:"fingerprint"`
+	FastOpen            bool     `yaml:"fast-open"`
+}
+
+// https://hysteria.network/docs/uri-scheme/
+// hysteria://host:port?protocol=udp&auth=123456&peer=sni.domain&insecure=1&upmbps=100&downmbps=100&alpn=hysteria&obfs=xplus&obfsParam=123456#remarks
+func hysteriaConf(body string) any {
+	u, err := url.Parse(body)
+	if err != nil {
+		log.Printf("parse hysteria failed, err: %v", err)
+		return nil
+	}
+
+	query := u.Query()
+	return &ClashHysteria{
+		Name:                u.Fragment,
+		Type:                "hysteria",
+		Server:              u.Hostname(),
+		Port:                cast.ToInt(u.Port()),
+		AuthStr:             query.Get("auth"),
+		Obfs:                query.Get("obfs"),
+		Alpn:                []string{query.Get("alpn")},
+		Protocol:            query.Get("protocol"),
+		Up:                  query.Get("upmbps"),
+		Down:                query.Get("downmbps"),
+		Sni:                 query.Get("peer"),
+		SkipCertVerify:      cast.ToBool(query.Get("insecure")),
+		RecvWindowConn:      cast.ToInt(query.Get("recv-window-conn")),
+		RecvWindow:          cast.ToInt(query.Get("recv-window")),
+		Ca:                  query.Get("ca"),
+		CaStr:               query.Get("ca-str"),
+		DisableMtuDiscovery: cast.ToBool(query.Get("disable_mtu_discovery")),
+		Fingerprint:         query.Get("fingerprint"),
+		FastOpen:            cast.ToBool(query.Get("fast-open")),
+	}
 }
 
 func v2rConf(s string) ClashVmess {
@@ -78,7 +154,7 @@ func v2rConf(s string) ClashVmess {
 	vmess := Vmess{}
 	err = json.Unmarshal(vmconfig, &vmess)
 	if err != nil {
-		log.Println(err)
+		log.Printf("v2ray config json unmarshal failed, err: %v", err)
 		return ClashVmess{}
 	}
 	clashVmess := ClashVmess{}
@@ -157,6 +233,7 @@ func ssrConf(s string) ClashRSSR {
 		return ClashRSSR{}
 	}
 	params := strings.Split(string(rawSSRConfig), `:`)
+
 	if len(params) != 6 {
 		return ClashRSSR{}
 	}
@@ -180,6 +257,7 @@ func ssrConf(s string) ClashRSSR {
 			ssr.Type = "ss"
 		}
 	}
+
 	suffix := strings.Split(params[SSRSuffix], "/?")
 	if len(suffix) != 2 {
 		return ClashRSSR{}
@@ -215,7 +293,6 @@ func ssrConf(s string) ClashRSSR {
 			continue
 		}
 	}
-
 	return ssr
 }
 
@@ -225,7 +302,7 @@ func ssConf(s string) ClashSS {
 		return ClashSS{}
 	}
 
-	findStr := regexp.MustCompile(`(?m)ss://([\/\+=\w]+)(@.+)?#(.+)`).FindStringSubmatch(s)
+	findStr := ssReg.FindStringSubmatch(s)
 	if len(findStr) < 4 {
 		return ClashSS{}
 	}
@@ -249,7 +326,7 @@ func ssConf(s string) ClashSS {
 	if findStr[5] != "" && strings.Contains(findStr[5], "plugin") {
 		query := findStr[5][strings.Index(findStr[5], "?")+1:]
 		queryMap, err := url.ParseQuery(query)
-		if err != nil{
+		if err != nil {
 			return ClashSS{}
 		}
 
@@ -259,19 +336,19 @@ func ssConf(s string) ClashSS {
 		case strings.Contains(ss.Plugin, "obfs"):
 			ss.Plugin = "obfs"
 			p.Mode = queryMap["obfs"][0]
-			if strings.Contains(query, "obfs-host="){
+			if strings.Contains(query, "obfs-host=") {
 				p.Host = queryMap["obfs-host"][0]
 			}
 		case ss.Plugin == "v2ray-plugin":
 			p.Mode = queryMap["mode"][0]
-			if strings.Contains(query, "host="){
+			if strings.Contains(query, "host=") {
 				p.Host = queryMap["host"][0]
 			}
-			if strings.Contains(query, "path="){
+			if strings.Contains(query, "path=") {
 				p.Path = queryMap["path"][0]
 			}
-			p.Mux = strings.Contains(query, "mux");
-			p.Tls = strings.Contains(query, "tls");
+			p.Mux = strings.Contains(query, "mux")
+			p.Tls = strings.Contains(query, "tls")
 			p.SkipCertVerify = true
 		}
 		ss.PluginOpts = p
@@ -310,17 +387,4 @@ func trojanConf(s string) Trojan {
 		Password: findStr[1],
 		Port:     findStr[3],
 	}
-}
-
-func filterNode(nodeName string) bool {
-	// 过滤剩余流量
-	if strings.Contains(nodeName, "剩余流量") {
-		return true
-	}
-
-	if strings.Contains(nodeName, "过期时间") {
-		return true
-	}
-
-	return false
 }
