@@ -10,31 +10,31 @@ const (
 )
 
 type Vmess struct {
-	Add  string      `json:"add"`
-	Aid  interface{} `json:"aid"`
-	Host string      `json:"host"`
-	ID   string      `json:"id"`
-	Net  string      `json:"net"`
-	Path string      `json:"path"`
-	Port interface{} `json:"port"`
-	PS   string      `json:"ps"`
-	TLS  string      `json:"tls"`
-	Type string      `json:"type"`
-	V    interface{} `json:"v"`
+	Add  string `json:"add"`
+	Aid  any    `json:"aid"`
+	Host string `json:"host"`
+	ID   string `json:"id"`
+	Net  string `json:"net"`
+	Path string `json:"path"`
+	Port any    `json:"port"`
+	PS   string `json:"ps"`
+	TLS  string `json:"tls"`
+	Type string `json:"type"`
+	V    any    `json:"v"`
 }
 
 type ClashVmess struct {
-	Name           string      `json:"name,omitempty"`
-	Type           string      `json:"type,omitempty"`
-	Server         string      `json:"server,omitempty"`
-	Port           interface{} `json:"port,omitempty"`
-	UUID           string      `json:"uuid,omitempty"`
-	AlterID        interface{} `json:"alterId,omitempty"`
-	Cipher         string      `json:"cipher,omitempty"`
-	TLS            bool        `json:"tls,omitempty"`
-	Network        string      `json:"network,omitempty"`
-	WSOpts         WSOptions   `json:"ws-opts,omitempty"`
-	SkipCertVerify bool        `json:"skip-cert-verify,omitempty"`
+	Name           string    `json:"name,omitempty"`
+	Type           string    `json:"type,omitempty"`
+	Server         string    `json:"server,omitempty"`
+	Port           any       `json:"port,omitempty"`
+	UUID           string    `json:"uuid,omitempty"`
+	AlterID        any       `json:"alterId,omitempty"`
+	Cipher         string    `json:"cipher,omitempty"`
+	TLS            bool      `json:"tls,omitempty"`
+	Network        string    `json:"network,omitempty"`
+	WSOpts         WSOptions `json:"ws-opts,omitempty"`
+	SkipCertVerify bool      `json:"skip-cert-verify,omitempty"`
 }
 
 type WSOptions struct {
@@ -45,36 +45,36 @@ type WSOptions struct {
 }
 
 type ClashRSSR struct {
-	Name          string      `json:"name"`
-	Type          string      `json:"type"`
-	Server        string      `json:"server"`
-	Port          interface{} `json:"port"`
-	Password      string      `json:"password"`
-	Cipher        string      `json:"cipher"`
-	Protocol      string      `json:"protocol"`
-	ProtocolParam string      `json:"protocol-param"`
-	OBFS          string      `json:"obfs"`
-	OBFSParam     string      `json:"obfs-param"`
+	Name          string `json:"name,omitempty"`
+	Type          string `json:"type,omitempty"`
+	Server        string `json:"server,omitempty"`
+	Port          any    `json:"port,omitempty"`
+	Password      string `json:"password,omitempty"`
+	Cipher        string `json:"cipher,omitempty"`
+	Protocol      string `json:"protocol,omitempty"`
+	ProtocolParam string `json:"protocol-param,omitempty"`
+	OBFS          string `json:"obfs,omitempty"`
+	OBFSParam     string `json:"obfs-param,omitempty"`
 }
 
 type ClashSS struct {
-	Name       string      `json:"name"`
-	Type       string      `json:"type"`
-	Server     string      `json:"server"`
-	Port       interface{} `json:"port"`
-	Password   string      `json:"password"`
-	Cipher     string      `json:"cipher"`
+	Name       string      `json:"name,omitempty"`
+	Type       string      `json:"type,omitempty"`
+	Server     string      `json:"server,omitempty"`
+	Port       any         `json:"port,omitempty"`
+	Password   string      `json:"password,omitempty"`
+	Cipher     string      `json:"cipher,omitempty"`
 	Plugin     string      `json:"plugin,omitempty"`
-	PluginOpts *PluginOpts  `json:"plugin-opts,omitempty"`
+	PluginOpts *PluginOpts `json:"plugin-opts,omitempty"`
 }
 
 type PluginOpts struct {
-	Mode string `json:"mode"`
-	Host string `json:"host,omitempty"`
-	Tls bool `json:"tls,omitempty"`
-	Path string `json:"path,omitempty"`
-	Mux bool `json:"mux,omitempty"`
-	SkipCertVerify bool `json:"skip-cert-verify,omitempty"`
+	Mode           string `json:"mode"`
+	Host           string `json:"host,omitempty"`
+	Tls            bool   `json:"tls,omitempty"`
+	Path           string `json:"path,omitempty"`
+	Mux            bool   `json:"mux,omitempty"`
+	SkipCertVerify bool   `json:"skip-cert-verify,omitempty"`
 }
 
 type SSD struct {
@@ -111,18 +111,22 @@ type Clash struct {
 	// ExternalUI         string                   `yaml:"external-ui"`
 	// Secret             string                   `yaml:"secret"`
 	// Experimental       map[string]interface{} 	`yaml:"experimental"`
-	Proxy             []map[string]interface{} `yaml:"proxies"`
-	ProxyGroup        []map[string]interface{} `yaml:"proxy-groups"`
-	Rule              []string                 `yaml:"rules"`
-	CFWByPass         []string                 `yaml:"cfw-bypass"`
-	CFWLatencyTimeout int                      `yaml:"cfw-latency-timeout"`
+	Proxies           []map[string]any `yaml:"proxies"`
+	ProxyGroups       []map[string]any `yaml:"proxy-groups"`
+	Rule              []string         `yaml:"rules"`
+	CFWByPass         []string         `yaml:"cfw-bypass"`
+	CFWLatencyTimeout int              `yaml:"cfw-latency-timeout"`
+
+	path       string
+	rawProxies []any
+	nodeOnly   bool
 }
 
 type Trojan struct {
-	Name     string      `json:"name"`
-	Type     string      `json:"type"`
-	Server   string      `json:"server"`
-	Password string      `json:"password"`
-	Sni      string      `json:"sni,omitempty"`
-	Port     interface{} `json:"port"`
+	Name     string `json:"name"`
+	Type     string `json:"type"`
+	Server   string `json:"server"`
+	Password string `json:"password"`
+	Sni      string `json:"sni,omitempty"`
+	Port     any    `json:"port"`
 }
