@@ -1,9 +1,9 @@
 package main
 
 import (
-	"github.com/whoisix/subscribe2clash/boot"
+	"github.com/icpd/subscribe2clash/boot"
 )
 
 func main() {
-	boot.Init()
+	boot.Run()
 }
